Update
Added burger transition to allow nav-bar to fill screen.
Troubleshooted why nav-header was not displaying on nav-bar screen. z-indexes were misaligned and position of nav-header had to adjusted to relative.
gsap was used to create rotation effect of navigation lines to create x and to change the color of the nav-header items black.