Update
Added transitioning effect using Barba library in JavaScript. Includes fade in and out, swipe color animations, and preserving header on both pages while swapping out the body for a seamless website experience.
Updated fashion page with the use of updated .html link in order to avoid having to click twice on the same page. A bug that is a product of using the Barba library. Keep in mind for future reference.